refactor(form-builder): move focus ring utils/hoc to components
