+++ packages/@sanity/form-builder/src/components/withFocusRing/helpers.ts
