+++ packages/@sanity/form-builder/src/components/withFocusRing/index.ts
@@ -1,0 +1,1 @@
+export * from './withFocusRing'
